✨ Cell Component 기능 CellItem Component로 이전
